requirements.txt
flask
flask-cors
python-dotenv
openai
selenium
webdriver-manager
requests
